% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{fit.dist}
\alias{fit.dist}
\title{Fit distribution to data}
\usage{
fit.dist(
  trait.data,
  trait = colnames(trait.data),
  dists = c("weibull", "lognormal", "gamma"),
  n = NULL
)
}
\arguments{
\item{trait.data}{data for distribution}

\item{trait}{name of trait to fit.
One of "tt", "sla", "rrr", "q"}

\item{dists}{list of distribution names}

\item{n}{only used in return value}
}
\value{
best fit distribution
}
\description{
Fit a distribution to data
}
\author{
David LeBauer
}
