% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.densities}
\alias{plot.densities}
\title{Plot Trait Probability Densities}
\usage{
\method{plot}{densities}(density.plot.inputs, outdir, ...)
}
\arguments{
\item{sensitivity.results}{list containing sa.samples and sa.splines}

\item{outdir}{}
}
\value{
outputs plots in outdir/sensitivity.analysis.pdf file
}
\description{
Plot probability density and data
}
\author{
David LeBauer
}

