% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.trait}
\alias{plot.trait}
\title{Plot trait density}
\usage{
plot.trait(trait, prior = NULL, posterior.sample = NULL, trait.df = NULL,
  fontsize = list(title = 18, axis = 14), x.lim = NULL, y.lim = NULL,
  logx = FALSE)
}
\arguments{
\item{trait}{character, name of trait to be plotted}

\item{prior}{named distribution with parameters}

\item{posterior.sample}{}

\item{fontsize}{}

\item{trait.data}{}
}
\value{
plot (grob) object
}
\description{
Plot trait density and data
}
\examples{
\dontrun{
prior1 <- data.frame(distn = 'norm',
                     parama = 20,
                     paramb = 5)
data1  <- data.frame(Y = c(19, 21), se = c(1,1))
plot.trait(trait = 'Vcmax',
          prior = prior1,
          trait.df = data1)
}
}
\author{
David LeBauer
}

