% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_densities}
\alias{plot_densities}
\title{Plot Trait Probability Densities}
\usage{
plot_densities(density.plot_inputs, outdir, ...)
}
\arguments{
\item{outdir}{directory in which to generate figure as pdf}

\item{sensitivity.results}{list containing sa.samples and sa.splines}
}
\value{
outputs plots in outdir/sensitivity.analysis.pdf file
}
\description{
Plot probability density and data
}
\author{
David LeBauer
}
