% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{pr.samp}
\alias{pr.samp}
\title{Sample from prior}
\usage{
pr.samp(distn, parama, paramb, n)
}
\arguments{
\item{distn}{name of distribution, e.g. "norm", "pois"}

\item{parama}{first parameter for distn call}

\item{paramb}{second parameter for distn call}

\item{n}{number of samples to return}
}
\value{
vector with n random samples from prior
}
\description{
Take n random samples from prior
}
\seealso{
\code{\link{get.sample}}
}
