% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{priorfig}
\alias{priorfig}
\title{Plot prior density and data}
\usage{
priorfig(
  priordata = NA,
  priordensity = NA,
  trait = NA,
  xlim = "auto",
  fontsize = 18
)
}
\arguments{
\item{priordata}{observations to be plotted as points}

\item{priordensity}{density of prior distribution, calculated by \code{prior.density}}

\item{trait}{dataframe with id, figid and units of the trait}

\item{xlim}{limits for x axis}

\item{fontsize}{passed to ggplot to set text size}
}
\value{
plot / grob of prior distribution with data used to inform the distribution
}
\description{
Plot prior density and data
}
\author{
David LeBauer
}
