% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{create.density.df}
\alias{create.density.df}
\title{Create Density Data Frame from Sample}
\usage{
create.density.df(samps = NULL, zero.bounded = FALSE, distribution = NULL,
  n = 1000, ...)
}
\arguments{
\item{samps}{a vector of samples from a distribution}

\item{zero.bounded}{}

\item{distribution}{list with elements distn, parama, paramb,
e.g. \code{list('norm', 0, 1)}}
}
\value{
data frame with x and y = dens(x)
}
\description{
Returns a data frame from \link{stats::density} function
}
\examples{
prior.df <- create.density.df(distribution = list('norm',0,1))
plot(prior.df)
samp.df <- create.density.df(samps = rnorm(100))
lines(samp.df)
}
\author{
David LeBauer
}

