% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get.sample}
\alias{get.sample}
\title{Get Samples}
\usage{
get.sample(prior, n, p = NULL)
}
\arguments{
\item{prior}{data.frame with distn, parama, paramb}

\item{n}{number of samples to return}

\item{p}{probability vector, pre-generated upstream to be used in the quantile function}
}
\value{
vector with n random samples from prior
}
\description{
Take n random samples from prior
}
\details{
Like pr.samp, with prior as a single input
}
\seealso{
\link{pr.samp}
}
