\name{plot.densities}
\alias{plot.densities}
\title{Plot Trait Probability Densities}
\usage{
  plot.densities(density.plot.inputs, outdir, ...)
}
\arguments{
  \item{sensitivity.results}{list containing sa.samples and
  sa.splines}

  \item{outdir}{}
}
\value{
  outputs plots in outdir/sensitivity.analysis.pdf file
}
\description{
  Plot probability density and data
}
\author{
  David LeBauer
}

