% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.densities}
\alias{plot.densities}
\title{Plot Trait Probability Densities}
\usage{
\method{plot}{densities}(density.plot.inputs, outdir, ...)
}
\arguments{
\item{sensitivity.results}{list containing sa.samples and sa.splines}
}
\value{
outputs plots in outdir/sensitivity.analysis.pdf file
}
\description{
Plot probability density and data
}
\author{
David LeBauer
}

