\name{pr.samp}
\alias{pr.samp}
\title{Sample from prior}
\usage{
  pr.samp(distn, parama, paramb, n)
}
\arguments{
  \item{distn}{}

  \item{parama}{}

  \item{paramb}{}

  \item{n}{number of samples to return}
}
\value{
  vector with n random samples from prior
}
\description{
  Take n random samples from prior
}
\seealso{
  \{code{\link{get.sample}}
}

